+++ src/Opulence/Framework/Authentication/Bootstrappers/AuthenticationBootstrapper.php
@@ -17,14 +17,13 @@ use Opulence\Authentication\Credentials\Authenticators\IAuthenticator;
 use Opulence\Authentication\Credentials\Authenticators\IAuthenticatorRegistry;
 use Opulence\Authentication\IAuthenticationContext;
 use Opulence\Authentication\Users\Orm\IUserRepository;
-use Opulence\Ioc\Bootstrappers\Bootstrapper;
-use Opulence\Ioc\Bootstrappers\ILazyBootstrapper;
+use Opulence\Ioc\Bootstrappers\LazyBootstrapper;
 use Opulence\Ioc\IContainer;
 
 /**
  * Defines the authentication bootstrapper
  */
-abstract class AuthenticationBootstrapper extends Bootstrapper implements ILazyBootstrapper
+abstract class AuthenticationBootstrapper extends LazyBootstrapper
 {
     /**
      * @inheritdoc